# improve performance (cargo bench) by 20-30%