The RootIterator is the heart of all detection logic. It performs up to tens or even hundreds of thousands of iterations per beat detection.

There is good potential to reduce the amount of needed iterations with only sacrificing minimal loss in accuracy - which is even irrelevant given the high sampling rates.


Benchmark `cargo bench` on my `11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz`:

**before**

```
simulate beat detection (with lowpass) with 4096 samples per invocation
                        time:   [1.9795 ms 1.9882 ms 1.9988 ms]
                        change: [+21.442% +22.009% +22.798%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe

simulate beat detection (no lowpass) with 4096 samples per invocation
                        time:   [2.0585 ms 2.0809 ms 2.1089 ms]
                        change: [+20.353% +21.636% +23.305%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe

```


**after**
```
simulate beat detection (with lowpass) with 4096 samples per invocation
                        time:   [1.6477 ms 1.6498 ms 1.6523 ms]
                        change: [-17.421% -16.968% -16.584%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  3 (3.00%) high severe

Benchmarking simulate beat detection (no lowpass) with 4096 samples per invocation: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.7s, enable flat sampling, or reduce sample count to 50.
simulate beat detection (no lowpass) with 4096 samples per invocation
                        time:   [1.7221 ms 1.7230 ms 1.7239 ms]
                        change: [-18.165% -17.062% -16.138%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  2 (2.00%) high severe
```