@phip1611 2024-05-04
On my Intel i5 10th Generation, the performance boost is +30%.

I think, this is the best I can get for now. Further performance improvement ideas that I have:

- Internally, scale down the signal to 1/100 of the sampling rate. I only need a sampling rate of 200Hz to detect beats of 20-80 Hz.
- drop envelope detection, use state machine-like approach to detect beat (should work pretty well, I think)